requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
psycopg2-binary>=2.9.9
requests>=2.31.0
pyyaml>=6.0.1
pytest>=8.0.0
python-dotenv>=1.0.0
